requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncmy==0.2.8
black==23.3.0
blinker==1.6.2
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
filelock==3.12.0
Flask==2.3.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.24.0
identify==2.5.24
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pre-commit==3.3.1
psutil==5.9.5
pydantic==1.10.7
PyJWT==2.7.0
PyMySQL==1.0.3
pymysql-pool==0.4.3
PyPika==0.48.9
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.11
tortoise-orm==0.19.3
typing_extensions==4.5.0
virtualenv==20.23.0
Werkzeug==2.3.4
